fix data type.

+++ backends/npu/tests/unittests/test_activation_op.py
@@ -99,7 +99,7 @@ class TestAtan(TestActivation):
 
     def test_out_name(self):
         with fluid.program_guard(fluid.Program()):
-            np_x = np.array([0.1])
+            np_x = np.array([0.1]).astype(np.float32)
             data = paddle.static.data(name="X", shape=[-1, 1])
             out = paddle.atan(data, name="Y")
             exe = fluid.Executor(self.place)
